Build using clang on Windows (#2180)

+++ src/server/ua_session.h
@@ -118,7 +118,7 @@ UA_Session_dequeuePublishReq(UA_Session *session);
                        "Connection %i | SecureChannel %i | Session %.*s | " MSG "%.0s", \
                        ((SESSION)->header.channel ?                     \
                         ((SESSION)->header.channel->connection ?        \
-                         (SESSION)->header.channel->connection->sockfd : 0) : 0), \
+                         (int)((SESSION)->header.channel->connection->sockfd) : 0) : 0), \
                        ((SESSION)->header.channel ?                     \
                         (SESSION)->header.channel->securityToken.channelId : 0), \
                        (int)idString.length, idString.data, __VA_ARGS__); \
